+++ QueryBuilder.Tests/UpdateTests.cs
@@ -9,23 +9,6 @@ namespace SqlKata.Tests
     public class UpdateTests : TestSupport
     {
         [Fact]
-        public void InsertObject()
-        {
-            var query = new Query("Table").AsInsert(new
-            {
-                Name = "The User",
-                Age = new DateTime(2018, 1, 1),
-            });
-
-            var c = Compile(query);
-
-            Assert.Equal("INSERT INTO [Table] ([Name], [Age]) VALUES ('The User', '2018-01-01')", c[EngineCodes.SqlServer]);
-
-
-            Assert.Equal("INSERT INTO \"TABLE\" (\"NAME\", \"AGE\") VALUES ('The User', '2018-01-01')", c[EngineCodes.Firebird]);
-        }
-
-        [Fact]
         public void UpdateObject()
         {
             var query = new Query("Table").AsUpdate(new
